requirements.txt
wheel==0.40.0
pytest==7.2.1
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==9.4.0
pdfminer.six==20221105
python-Levenshtein
scikit-learn